refactor policy bundles code into more digestable chunks
